rename package to macos
